Fix 'accidental' module cycles touching scc1
